Heavy refactor that group executions in a key for each given run.
This implies sending the serialized execution to the nodes instead of
just the job.

This allows to control each "execution" in all nodes knowing wich ones
relates to each given cron execution.

It also carries the execution and the job in network messages and save it serialized to the store with each execution.