#184 Hotfix:
Panic is thrown on store.go:243. It is accessing to a store.Store member of the type Store which is nil.

This fix is doing the nil checking to return an error if client is not found.

I have also take advantage of this PR to fix a panic raised in tests. It is generally not recommended and much more idiomatic to use t.Fatal in a test.