Add retries for rpcconfig query timeouts
This should fix an issue that I have had very similar
to #237 where tasks would fail to register their results
because they couldn't get the RPC address of a server. This
coupled with a job config of "concurrency": "forbid" meant
jobs would stop indefinitely.

The path where rpcAddr is nil from queryRPCConfig() is
a path dkron should never return because it means that
results are never recorded.