Slightly improve processing of last execution group
This helps to speed up the UI some as the /dashboard/jobs
endpoint needs to find the last execution group of every
job to display some UI components. This removes the call
to s.GetExecuionGroup becuase we already have that information
so there is no need to recursively get all execution groups
from the KV Store again and increase time and resource
utilization.

This also adds a test for GetLastExecutionGroup to verify
intended outputs.

Converted store_test to use testify.