Improve API call on create or edit jobs
When creating jobs, a bunch of time is invested in sending the scheduler restart query and waiting for the response.

This sets refactor performs all store operations but runs the restart scheduler call in a new thread when a timeout is finished so it could be called multiple times and just the last one will be executed to reload jobs.