Fix dial tcp: missing address
Under some circunstances, network or other, getting the RPC address of a Dkron serve in a cluster via a Serf query can return an empty value making the agent unable to respond with the result.

In this PR the RPC address config of a server is embedded in the agent tags, making this config available to all members of the cluster in realtime. This makes innecessary to query the cluster for RPC config, and uses already present info in the agent.

This improves performance by saving network requests and ensures the fix to work under any condition.