Reflect store status in API
Dkron could be running but the storage engine could be unreachable.

We introduce a new return param in the status API path / that provides that info.

backend_status Will return the backend health status encoded in HTTP codes.

At the same time the endpoint will return the corresponding HTTP code.