Remove unnecessary updates of job finish times
agent.RefreshJobStatus() was added to prevent missed exeuctionDone
messages from blocking forbidden concurrency jobs. This checked all
executions to see if they were done before responding to the job status.
If the jobs were already recorded as finished, this is pointless and
causes a lot of extra overhead on the network for no gain. The
extra effort also caused the execution FinishedAt parameter to
be set to a value that is not accurate.