Mitigate bug of job stuck running
I'm not sure what caused this bug for me, but from the API, I
could query /v1/jobs and all statuses were "success" but when
the scheduler would try and run the job, it would log that it was
skipping execution because concurrency was forbidden and the job
was running. This updates what status is used for a job and instead of
using the status on the pointer, querries the status instead to have
a real time view of the job.