Add last run to cron timing logic
Previously, if dkron was down during the window that a job should run
and then started back up, the job would be scheduled at the next
interval or schedule from that point as if the job had just run.

This changes that functionality to schedule jobs based on their
last run time instead of the current time so that on startup,
missed jobs will run.