Add job name tags to log events to improve debugging
Some of the log calls had the job_name tag, but others did not. This adds them to those calls that were missing to improve log correlation, especially when there are many jobs all with the same schedule (i.e. @every 1m) and these execution log entries are all intertwined.