Improve tests for http executor
While troubleshooting issues with HTTP jobs, I noticed that test cases that I expected to fail, were not causing the test to fail. I dug in and realilzed that the errors from the output were not being checked. The returned error was being checked, but this is hardcoded to nil in the return so it offers very little value.

This adds some common test cases for the HTTP executor and ensures that many input validations return an error and known good configurations work as expected.

This also updates the Dockerfile to go 1.16 because 1.15 does not include go:embed and was causing `make test` to fail.